+++ aws-android-sdk-pinpoint/src/main/java/com/amazonaws/mobileconnectors/pinpoint/internal/core/PinpointContext.java
@@ -69,32 +69,20 @@ public class PinpointContext implements Serializable {
 
     public PinpointContext(final AmazonPinpointAnalyticsClient analyticsServiceClient,
             final AmazonPinpointClient pinpointServiceClient,
-            final Context applicationContext, final Regions region, final String appId,
+            final Context applicationContext, final String appId,
             final SDKInfo sdkInfo, final PinpointConfiguration pinpointConfiguration) {
         this.sdkInfo = sdkInfo;
-
         this.pinpointConfiguration = pinpointConfiguration;
         this.system = new AndroidSystem(applicationContext, appId);
-
         this.uniqueIdService = new SharedPrefsUniqueIdService(appId, applicationContext);
         this.uniqueId = uniqueIdService.getUniqueId(this);
-
-        analyticsServiceClient.setServiceNameIntern("mobileanalytics");
-        analyticsServiceClient.setRegion(Region.getRegion(region));
         this.analyticsServiceClient = analyticsServiceClient;
-
-        pinpointServiceClient.setServiceNameIntern("mobiletargeting");
-        pinpointServiceClient.setRegion(Region.getRegion(region));
-
         this.pinpointServiceClient = pinpointServiceClient;
+        this.applicationContext = applicationContext;
+        this.configuration = AndroidPreferencesConfiguration.newInstance(this);
 
         analyticsServiceClient.addRequestHandler(new SDKInfoHandler(sdkInfo));
         pinpointServiceClient.addRequestHandler(new SDKInfoHandler(sdkInfo));
-
-        // config settings was always a new HashMap
-        configuration = AndroidPreferencesConfiguration.newInstance(this);
-
-        this.applicationContext = applicationContext;
     }
 
     public NotificationClient getNotificationClient() {