+++ aws-android-sdk-rekognition/src/main/java/com/amazonaws/services/rekognition/model/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Classes modeling the various types represented by AmazonRekognition
+ */
+
+package com.amazonaws.services.rekognition.model;
+