+++ aws-android-sdk-rekognition/src/main/java/com/amazonaws/services/rekognition/model/transform/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Marhsalling for the various types represented byAmazonRekognition
+ */
+
+package com.amazonaws.services.rekognition.model.transform;
+